# chore: use enhanced `cmd-docgen` script for generated component documentation