--- a/package.json
+++ b/package.json
@@ -58,7 +58,7 @@
     "@zendeskgarden/react-theming": "8.42.2",
     "@zendeskgarden/react-tooltips": "8.42.2",
     "@zendeskgarden/react-typography": "8.42.2",
-    "@zendeskgarden/scripts": "1.2.1",
+    "@zendeskgarden/scripts": "1.3.0",
     "@zendeskgarden/stylelint-config": "16.0.1",
     "@zendeskgarden/svg-icons": "6.30.2",
     "babel-eslint": "10.1.0",