## Description

Updates `@zendeskgarden/scripts` dependency to v1.3.0 in order to take advantage of new `cmdDocgen` script. Enhanced caching is leveraged to achieve the build results shown below.

## Detail

Timing for `yarn-build`:

**Before**
- clean: 190s (70s to run page queries)
- cached: 94s (67s to run page queries)

**After**
- clean: 162s (37s to run page queries)
- cached: 29s (less than 1s to run page queries) 👈